refactor: exteact <ValueDisplay> to next-common & do polish, #1363
