+++ packages/kintsugi-next/utils/index.js
@@ -135,28 +135,3 @@ export function getTimelineStatus(type, method) {
   }
   return { value: method };
 }
-
-export function abbreviateBigNumber(x, fixed = 2) {
-  const n = new BigNumber(x);
-  const fmt = {
-    decimalSeparator: ".",
-    groupSeparator: ",",
-    groupSize: 3,
-  };
-  let divideBy = new BigNumber("1");
-  const bigNumbers = [
-    { bigNumber: new BigNumber("1000"), abbr: "K" },
-    { bigNumber: new BigNumber("1000000"), abbr: "M" },
-    { bigNumber: new BigNumber("1000000000"), abbr: "B" },
-    { bigNumber: new BigNumber("1000000000000"), abbr: "T" },
-    { bigNumber: new BigNumber("1000000000000000"), abbr: "Q" },
-  ];
-  bigNumbers.forEach((data) => {
-    if (n.isGreaterThan(data.bigNumber)) {
-      divideBy = data.bigNumber;
-      fmt.suffix = data.abbr;
-    }
-  });
-  BigNumber.config({ FORMAT: fmt });
-  return new BigNumber(n.dividedBy(divideBy).toFixed(fixed)).toFormat();
-}
