+++ packages/next-common/utils/viewfuncs/index.js
@@ -1,6 +1,47 @@
+import BigNumber from "bignumber.js";
+
 export function toApiType(type) {
   if (type === "treasury/bounty") {
     return "treasury/bounties";
   }
   return `${type}s`;
+}
+
+export function getEffectiveNumbers(n) {
+  const result = [];
+  let flag = false;
+  n.toString().split("")
+    .reverse()
+    .forEach((dig) => {
+      if (!isNaN(parseInt(dig))) {
+        flag = flag || parseInt(dig) > 0;
+        flag && result.push(dig);
+      }
+    });
+  return result.reverse().join();
+}
+
+export function abbreviateBigNumber(x, fixed = 2) {
+  const n = new BigNumber(x);
+  const fmt = {
+    decimalSeparator: ".",
+    groupSeparator: ",",
+    groupSize: 3,
+  };
+  let divideBy = new BigNumber("1");
+  const bigNumbers = [
+    { bigNumber: new BigNumber("1000"), abbr: "K" },
+    { bigNumber: new BigNumber("1000000"), abbr: "M" },
+    { bigNumber: new BigNumber("1000000000"), abbr: "B" },
+    { bigNumber: new BigNumber("1000000000000"), abbr: "T" },
+    { bigNumber: new BigNumber("1000000000000000"), abbr: "Q" },
+  ];
+  bigNumbers.forEach((data) => {
+    if (n.isGreaterThan(data.bigNumber)) {
+      divideBy = data.bigNumber;
+      fmt.suffix = data.abbr;
+    }
+  });
+  BigNumber.config({ FORMAT: fmt });
+  return new BigNumber(n.dividedBy(divideBy).toFixed(fixed)).toFormat();
 }
