+++ packages/next/components/referenda/popup/splitVoteStatus.js
@@ -7,7 +7,7 @@ import {
   WarningWrapper,
 } from "./styled";
 import Tooltip from "next-common/components/tooltip";
-import DisplayValue from "../displayValue";
+import DisplayValue from "next-common/components/displayValue";
 import ApproveIcon from "next-common/assets/imgs/icons/approve.svg";
 import RejectIcon from "next-common/assets/imgs/icons/reject.svg";
 import { toPrecision } from "utils";
