+++ packages/next/components/referenda/popup/standardVoteStatus.js
@@ -6,7 +6,7 @@ import {
   StatusWrapper,
   WarningWrapper,
 } from "./styled";
-import DisplayValue from "../displayValue";
+import DisplayValue from "next-common/components/displayValue";
 import ApproveIcon from "next-common/assets/imgs/icons/approve.svg";
 import RejectIcon from "next-common/assets/imgs/icons/reject.svg";
 import { toPrecision } from "utils";
