+++ packages/next/components/referenda/vote.js
@@ -19,7 +19,7 @@ import NayIcon from "public/imgs/icons/nay.svg";
 import TurnoutIcon from "public/imgs/icons/turnout.svg";
 import ElectorateIcon from "public/imgs/icons/electorate.svg";
 import Threshold from "./threshold";
-import DisplayValue from "./displayValue";
+import DisplayValue from "next-common/components/displayValue";
 import Loading from "./loading";
 import { useBestNumber } from "next-common/utils/hooks";
